Multipart Line Extension Bug Fix and Process Refactoring
Changed the way that road lines are extended to force them to be single
part before the extension takes place. This fixes an error wherein one
or both of the end points of two or more lines are selected as the
extension points. This would previously cause the extended line to not
reach the county buffer for intersection, resulting in no county road
name rosette point to be generated for the non-intersecting road line.
Also improved the code's readability by breaking some of the longer
functions into smaller functions.